Unit test for Response to Claim section saved into a separate file
